+++ src/Symfony/Component/Console/SignalRegistry/SignalRegistry.php
@@ -13,26 +13,27 @@ namespace Symfony\Component\Console\SignalRegistry;
 
 final class SignalRegistry
 {
-    private $registeredSignals = [];
-
-    private $handlingSignals = [];
+    private $signalHandlers = [];
 
     public function __construct()
     {
-        pcntl_async_signals(true);
+        if (\function_exists('pcntl_async_signals')) {
+            pcntl_async_signals(true);
+        }
     }
 
     public function register(int $signal, callable $signalHandler): void
     {
-        if (!isset($this->registeredSignals[$signal])) {
+        if (!isset($this->signalHandlers[$signal])) {
             $previousCallback = pcntl_signal_get_handler($signal);
 
             if (\is_callable($previousCallback)) {
-                $this->registeredSignals[$signal][] = $previousCallback;
+                $this->signalHandlers[$signal][] = $previousCallback;
             }
         }
 
-        $this->registeredSignals[$signal][] = $signalHandler;
+        $this->signalHandlers[$signal][] = $signalHandler;
+
         pcntl_signal($signal, [$this, 'handle']);
     }
 
@@ -41,20 +42,11 @@ final class SignalRegistry
      */
     public function handle(int $signal): void
     {
-        foreach ($this->registeredSignals[$signal] as $signalHandler) {
-            $signalHandler($signal);
-        }
-    }
+        $count = \count($this->signalHandlers[$signal]);
 
-    public function addHandlingSignals(int ...$signals): void
-    {
-        foreach ($signals as $signal) {
-            $this->handlingSignals[$signal] = true;
+        foreach ($this->signalHandlers[$signal] as $i => $signalHandler) {
+            $hasNext = $i !== $count - 1;
+            $signalHandler($signal, $hasNext);
         }
-    }
-
-    public function getHandlingSignals(): array
-    {
-        return array_keys($this->handlingSignals);
     }
 }